Extract TraktLookup to own file
